+++ src/components/internal-link/internal-link.js
@@ -5,9 +5,9 @@ import { Link } from 'react-router-dom';
 /**
  *
  * This Link component is used to track the origination
- * of a link wether its from within the application or not.
+ * of a link whether it's from within the application or not.
  * Here the Link Component sets the state of eholdings to true. Checks
- * that 'to' is typeof string. Adds a property 'pathname' to location has
+ * that 'to' is typeof string. Adds a property 'pathname' to location
  * containing the 'pathname' that was passed in. Then is returns a new
  * <Link to={ location } {...props } /> component
  * with a location hash containing keys of pathname and state.
@@ -19,11 +19,11 @@ import { Link } from 'react-router-dom';
  *
  * --- end Usage ---
  *
- * With this added data you can test wether a user
+ * With this added data you can test whether a user
  * navigated to a section of the app via the use of links vs
  * pasting in URL to browser using the router history.
  *
- * ie. In comsuming component you can check router history.
+ * ie. In consuming component you can check router history.
  * let historyState = router.history.location.state;
  *
  * If you look at historyState it would be
@@ -34,10 +34,10 @@ import { Link } from 'react-router-dom';
  *
  */
 
-export default function InternalLink({
+const InternalLink = ({
   to,
   ...props
-}) {
+}) => {
   let location = { state: { eholdings: true } };
 
   if (typeof to === 'string') {
@@ -49,8 +49,15 @@ export default function InternalLink({
   return (
     <Link to={location} {...props} />
   );
-}
+};
 
 InternalLink.propTypes = {
-  to: PropTypes.oneOfType([PropTypes.string, PropTypes.object])
+  to: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.shape({
+      pathname: PropTypes.string,
+    }),
+  ]),
 };
+
+export default InternalLink;